Tests/LibWeb/Text/input/css/transition-behavior-multiple.html

<!DOCTYPE html>
<html>
    <style>
        #foo {
            transition-behavior: allow-discrete, normal;
        }
    </style>
    <script src="../include.js"></script>
    <script>
        test(() => {
            println(document.styleSheets[0].cssRules[0].cssText);
        });
    </script>
</html>
