Tests/LibWeb/Text/expected/css/transition-behavior-multiple.txt

#foo { transition-behavior: allow-discrete, normal; }
